refactor: separated components from header
