+++ libs/shared/src/lib/classes/models/snackbar.ts
@@ -1,0 +1,12 @@
+import { AlertColor } from "@mui/material";
+
+export interface BytebankSnackbarProps {
+    open: boolean; 
+    onClose: () => void;
+    data: SnackbarData | null;
+}
+
+export interface SnackbarData {
+    status: AlertColor; 
+    message?: string;
+}
