+++ libs/shared/src/lib/components/header/index.tsx
@@ -3,13 +3,10 @@ import './style.scss';
 
 import { useRouter } from 'next/navigation';
 import {
-  Alert,
-  AlertColor,
   AppBar,
   Box,
   Container,
   Link,
-  Snackbar,
   MenuItem,
   Typography,
   Menu,
@@ -25,38 +22,33 @@ import AppRegistrationRoundedIcon from '@mui/icons-material/AppRegistrationRound
 import { BytebankMenu } from '../menu';
 import { BytebankButton } from '../button';
 import { ReactElement, useState } from 'react';
-import { BytebankModal } from '../modal';
-import { FormProvider, useForm } from 'react-hook-form';
-import { BytebankInputController } from '../input/ControlledInput';
-import { User } from '../../classes/models/user';
 import { BytebankText } from '../text';
 import {
   loggedRoutes,
   unloggedRoutes,
 } from '../../classes/constants/routes.config';
 import { useUser } from '../../contexts/user.context';
+import { BytebankLoginModal } from '../../modals/login-modal';
+import { BytebankRegisterModal } from '../../modals/register-modal';
+import { BytebankSnackbar } from '../snackbar';
+import { SnackbarData } from '../../classes/models/snackbar';
 
 interface HeaderProps {
   mobile?: boolean;
 }
 
-export function BytebankHeader({ mobile }: HeaderProps) {
+export function BytebankHeader({ mobile }: HeaderProps): ReactElement {
   const router = useRouter();
   const theme = useTheme<Theme>();
   const isMobile = useMediaQuery(theme.breakpoints.down('md'));
 
   const [anchorElUser, setAnchorElUser] = useState<null | HTMLElement>(null);
-  const [snackbarData, setSnackbarData] = useState<{
-    severity: AlertColor;
-    message: string;
-  } | null>(null);
+  const [snackbarData, setSnackbarData] = useState<SnackbarData | null>(null);
   const [isSnackbarOpen, setSnackbarOpen] = useState(false);
   const [openLoginModal, setOpenLoginModal] = useState(false);
-  const [isLoginLoading, setLoginLoading] = useState(false);
-  const [isRegisterLoading, setRegisterLoading] = useState(false);
   const [openRegisterModal, setOpenRegisterModal] = useState(false);
   const { user, setUser, loading } = useUser();
-  
+
   const isLogged = !!user;
   const settings = [
     {
@@ -69,21 +61,6 @@ export function BytebankHeader({ mobile }: HeaderProps) {
     { name: 'Sair', action: () => handleLogout() },
   ];
 
-  const loginMethods = useForm<{ email: string; password: string }>({
-    defaultValues: {
-      email: '',
-      password: '',
-    },
-  });
-
-  const registerMethods = useForm<Partial<User>>({
-    defaultValues: {
-      name: '',
-      email: '',
-      password: '',
-    },
-  });
-
   const redirectTo = (name: string) => router.push(name);
 
   const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => setAnchorElUser(event.currentTarget);
@@ -91,14 +68,32 @@ export function BytebankHeader({ mobile }: HeaderProps) {
   const closeLoginModal = () => setOpenLoginModal(false);
   const closeRegisterModal = () => setOpenRegisterModal(false);
 
-  const handleLoginModal = () => {
-    closeRegisterModal();
-    setOpenLoginModal(true);
+  const closeSnackbar = () => {
+    setSnackbarOpen(false);
+    setSnackbarData(null)
   };
 
-  const handleRegisterModal = () => {
+  const handleLoginModal = ({ status, message }: SnackbarData) => {
+    if (status === 'success') {
+      router.push('/home');
+      setTimeout(() => {
+        closeLoginModal();
+      }, 20000);
+    } else {
+      closeRegisterModal();
+      setSnackbarData({ status, message });
+      setSnackbarOpen(true);
+    }
+  };
+
+  const handleRegisterModal = ({ status, message }: SnackbarData) => {
+    if (status === 'success') {
+      closeRegisterModal();
+    }
+
     closeLoginModal();
-    setOpenRegisterModal(true);
+    setSnackbarData({ status, message });
+    setSnackbarOpen(true);
   };
 
   const handleLogout = (): void => {
@@ -106,206 +101,6 @@ export function BytebankHeader({ mobile }: HeaderProps) {
     setUser(null);
     redirectTo('/');
   };
-
-  const handleRegister = async (data: Partial<User>) => {
-    setRegisterLoading(true);
-    const response = await fetch('/api/register', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (response.ok) {
-      const res = (await response.json()) as { message: string };
-      setSnackbarData({ severity: 'success', message: res.message });
-      registerMethods.reset();
-      closeRegisterModal();
-    } else {
-      const responseError = (await response.json()) as { error: string };
-      setSnackbarData({ severity: 'error', message: responseError.error });
-    }
-    setRegisterLoading(false);
-    setSnackbarOpen(true);
-  };
-
-  const handleLogin = async (data: { email: string; password: string }) => {
-    setLoginLoading(true);
-    const response = await fetch('/api/login', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (response.ok) {
-      const userData = (await response.json()) as User;
-      setUser(userData);
-      loginMethods.reset();
-
-      router.push('/home'); // <-- redirect para /home
-
-      setTimeout(() => {
-        closeLoginModal(); 
-        setLoginLoading(false);
-      }, 20000); 
-
-    } else {
-      const responseError = (await response.json()) as { error: string };
-      setSnackbarData({ severity: 'error', message: responseError.error });
-      setLoginLoading(false);
-    }
-    setLoginLoading(false);
-    setSnackbarOpen(true);
-  };
-
-  const renderSnackbar = () => {
-    const handleSnackbarClose = () => {
-      setSnackbarOpen(false);
-      setSnackbarData(null);
-    };
-
-    return snackbarData ? (
-      <>
-        <Snackbar
-          open={isSnackbarOpen}
-          autoHideDuration={6000}
-          onClose={handleSnackbarClose}
-          anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-        >
-          <Alert onClose={handleSnackbarClose} severity={snackbarData.severity}>
-            {snackbarData.message}
-          </Alert>
-        </Snackbar>
-      </>
-    ) : null;
-  };
-
-  const renderRegisterModal = (isOpen: boolean): ReactElement => (
-    <>
-      <BytebankModal
-        title={'Criar uma conta'}
-        illustration={'register'}
-        illustrationSize={'lg'}
-        open={isOpen}
-        onClose={() => closeRegisterModal()}
-      >
-        <>
-          <BytebankText>
-            Preencha os campos abaixo para criar sua conta corrente!
-          </BytebankText>
-          <FormProvider {...registerMethods}>
-            <form onSubmit={registerMethods.handleSubmit(handleRegister)}>
-              <BytebankInputController
-                name="name"
-                autoComplete="name"
-                type="text"
-                label="Nome"
-                placeholder="Digite seu nome"
-              />
-              <BytebankInputController
-                name="email"
-                autoComplete="email"
-                type="email"
-                label="E-mail"
-                placeholder="Digite seu e-mail"
-              />
-              <BytebankInputController
-                name="password"
-                autoComplete="new-password"
-                type="password"
-                label="Senha"
-                placeholder="Digite sua senha"
-              />
-              <Box display={'flex'} pt={2} justifyContent={'center'}>
-                <BytebankButton
-                  label={'Criar conta'}
-                  color={'secondary'}
-                  variant={'contained'}
-                  loading={isRegisterLoading}
-                  fullWidth
-                ></BytebankButton>
-              </Box>
-            </form>
-          </FormProvider>
-          <Box pt={4} display={'flex'} gap={1} justifyContent={'center'}>
-            <BytebankText>Já tem uma conta?</BytebankText>
-            <Link
-              component="button"
-              variant="sm"
-              color={'secondary'}
-              onClick={handleLoginModal}
-            >
-              Fazer login
-            </Link>
-          </Box>
-        </>
-      </BytebankModal>
-    </>
-  );
-
-  const renderLoginModal = (isOpen: boolean): ReactElement => (
-    <>
-      <BytebankModal
-        title={'Login'}
-        illustration={'login'}
-        illustrationSize={'lg'}
-        open={isOpen}
-        onClose={() => closeLoginModal()}
-      >
-        <>
-          <FormProvider {...loginMethods}>
-            <form onSubmit={loginMethods.handleSubmit(handleLogin)}>
-              <BytebankInputController
-                name="email"
-                autoComplete="email"
-                type="email"
-                label="E-mail"
-                placeholder="Digite seu e-mail"
-              />
-              <BytebankInputController
-                name="password"
-                autoComplete="current-password"
-                type="password"
-                label="Senha"
-                placeholder="Digite sua senha"
-              />
-              <Box
-                display={'flex'}
-                gap={2}
-                flexDirection={'column'}
-                justifyContent={'center'}
-              >
-                <Link component="button" variant="sm" color={'secondary'}>
-                  Esqueceu sua senha?
-                </Link>
-                <BytebankButton
-                  label={'Entrar'}
-                  color={'secondary'}
-                  variant={'contained'}
-                  loading={isLoginLoading}
-                  fullWidth
-                ></BytebankButton>
-              </Box>
-            </form>
-          </FormProvider>
-          <Box pt={4} display={'flex'} gap={1} justifyContent={'center'}>
-            <BytebankText>Não tem uma conta?</BytebankText>{' '}
-            <Link
-              component="button"
-              variant="sm"
-              color={'secondary'}
-              onClick={handleRegisterModal}
-            >
-              Crie uma agora!
-            </Link>
-          </Box>
-        </>
-      </BytebankModal>
-    </>
-  );
 
   const renderMenuSettings = () => {
     return (
@@ -388,13 +183,13 @@ export function BytebankHeader({ mobile }: HeaderProps) {
                 !isMobile ? (
                   <Box display={'flex'} flex={'none'} gap={2}>
                     <BytebankButton
-                      sendSubmit={handleRegisterModal}
+                      sendSubmit={() => setOpenRegisterModal(true)}
                       label="Crie uma conta"
                       color="success"
                       variant="contained"
                     />
                     <BytebankButton
-                      sendSubmit={handleLoginModal}
+                      sendSubmit={() => setOpenLoginModal(true)}
                       label="Entre"
                       color="success"
                       variant="outlined"
@@ -417,9 +212,9 @@ export function BytebankHeader({ mobile }: HeaderProps) {
         </AppBar>
       )}
 
-      {renderRegisterModal(openRegisterModal)}
-      {renderLoginModal(openLoginModal)}
-      {renderSnackbar()}
+      <BytebankRegisterModal open={openRegisterModal} onClose={closeRegisterModal} onSubmit={handleRegisterModal} openLoginModal={() => { closeRegisterModal(); setOpenLoginModal(true) }} />
+      <BytebankLoginModal open={openLoginModal} onClose={closeLoginModal} onSubmit={handleLoginModal} openRegisterModal={() => { closeLoginModal(); setOpenRegisterModal(true) }} />
+      <BytebankSnackbar open={isSnackbarOpen} data={snackbarData} onClose={closeSnackbar} />
     </>
   );
 }
