+++ libs/shared/src/lib/modals/login-modal/index.tsx
@@ -1,0 +1,108 @@
+import { Box, Link } from "@mui/material";
+import { ReactElement, useState } from "react";
+import { useForm, FormProvider } from "react-hook-form";
+import { useUser, User, BytebankModal, BytebankInputController, BytebankButton, BytebankText } from "../../shared";
+import { SnackbarData } from "../../classes/models/snackbar";
+
+interface LoginModalProps {
+    open: boolean;
+    onClose: () => void;
+    onSubmit: (data: SnackbarData) => void;
+    openRegisterModal: () => void;
+}
+
+export function BytebankLoginModal({ open, onClose, onSubmit, openRegisterModal }: LoginModalProps): ReactElement {
+    const [isLoading, setLoading] = useState(false);
+    const { setUser } = useUser();
+
+    const loginMethods = useForm<{ email: string; password: string }>({
+        defaultValues: {
+            email: '',
+            password: '',
+        },
+    });
+
+    const handleLogin = async (data: { email: string; password: string }) => {
+        setLoading(true);
+        const response = await fetch('/api/login', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify(data),
+        });
+
+        if (response.ok) {
+            const userData = (await response.json()) as User;
+            setUser(userData);
+            loginMethods.reset();
+            onSubmit({ status: 'success'});
+        } else {
+            const responseError = (await response.json()) as { error: string };
+            onSubmit({ status: 'error', message: responseError.error });
+            setLoading(false);
+        }
+        setLoading(false);
+    };
+
+    return (
+        <>
+            <BytebankModal
+                title={'Login'}
+                illustration={'login'}
+                illustrationSize={'lg'}
+                open={open}
+                onClose={() => onClose()}
+            >
+                <>
+                    <FormProvider {...loginMethods}>
+                        <form onSubmit={loginMethods.handleSubmit(handleLogin)}>
+                            <BytebankInputController
+                                name="email"
+                                autoComplete="email"
+                                type="email"
+                                label="E-mail"
+                                placeholder="Digite seu e-mail"
+                            />
+                            <BytebankInputController
+                                name="password"
+                                autoComplete="current-password"
+                                type="password"
+                                label="Senha"
+                                placeholder="Digite sua senha"
+                            />
+                            <Box
+                                display={'flex'}
+                                gap={2}
+                                flexDirection={'column'}
+                                justifyContent={'center'}
+                            >
+                                <Link component="button" variant="sm" color={'secondary'}>
+                                    Esqueceu sua senha?
+                                </Link>
+                                <BytebankButton
+                                    label={'Entrar'}
+                                    color={'secondary'}
+                                    variant={'contained'}
+                                    loading={isLoading}
+                                    fullWidth
+                                ></BytebankButton>
+                            </Box>
+                        </form>
+                    </FormProvider>
+                    <Box pt={4} display={'flex'} gap={1} justifyContent={'center'}>
+                        <BytebankText>Não tem uma conta?</BytebankText>{' '}
+                        <Link
+                            component="button"
+                            variant="sm"
+                            color={'secondary'}
+                            onClick={openRegisterModal}
+                        >
+                            Crie uma agora!
+                        </Link>
+                    </Box>
+                </>
+            </BytebankModal>
+        </>
+    );
+}
